* Added support for MXML declaration of RemoteObjectStub results and faults. Especially usefully for use with DI frameworks such as Parsley or Swiz.
*Allow to ignore arguments (ignoreArgs=true) when setting a result or a fault for a given operation.
This will return the result or fault for any arguments passed to the operation.
This feature is useful for SUT, where results/faults are set before each test,
so we don't care about the arguments

*Allow to partially stub a RemoteObject: all operations for which a result or a fault have not be set will
perform a real RPC call.
This is useful if you just want to stub some operation but not all (eg. stub the login )

Important Note: these modifications were made prior to addition of argument signature,
so merge of both modifications needs to be done.